+++ core/producer/stage.h
@@ -72,8 +72,8 @@ public:
 	void clear(int index);
 	void clear();	
 	void swap_layers(const safe_ptr<stage>& other);
-	void swap_layer(int index, size_t other_index);
-	void swap_layer(int index, size_t other_index, const safe_ptr<stage>& other);
+	void swap_layer(int index, uint32_t other_index);
+	void swap_layer(int index, uint32_t other_index, const safe_ptr<stage>& other);
 
 	void add_layer_consumer(void* token, int layer, const std::shared_ptr<write_frame_consumer>& layer_consumer);
 	void remove_layer_consumer(void* token, int layer);
